refactor: move aside out of layout
