+++ packages/mirror-tv-next/styles/components/category/layout/aside.module.scss
@@ -1,0 +1,21 @@
+@import '/styles/theme/breakpoints';
+
+.aside {
+  display: flex;
+  flex-direction: column;
+  @include media-breakpoint-up(md) {
+    flex-direction: row;
+    width: 688px;
+    display: flex;
+    align-items: stretch;
+    justify-content: space-between;
+    gap: 16px;
+  }
+  @include media-breakpoint-up(xl) {
+    flex-direction: column;
+    width: 384px;
+    display: block;
+    padding: 0;
+    margin-top: 60px;
+  }
+}
